\name{macro}
\alias{macro}
\docType{data}
\title{Economic Indicators of 20 OECD countries for 1999}
\description{Data on the macroeconomic performance of national economies of 20 countries, members of the OECD (September 1999). The performance of the economies reflects the interaction of six main economic indicators (percentage change from the previous year):  gross domestic product (GDP), leading indicator (LI), unemployment rate (UR), interest rate (IR), trade balance (TB), net national savings (NNS).}
\usage{data(macro)}
\format{
  A data frame with 20 observations on the following 6 variables.
  \describe{
    \item{\code{GDP}}{numeric}
    \item{\code{LI}}{numeric}
    \item{\code{UR}}{numeric}
    \item{\code{IR}}{numeric}
    \item{\code{TB}}{numeric}
    \item{\code{NNS}}{numeric}
  }
}
\source{Vichi, M. & Kiers, H. A. (2001). Factorial k-means analysis for two-way data. \emph{Computational Statistics & Data Analysis}, \emph{37}(1), 49-64.}
\keyword{datasets}
