\name{underwear}
\alias{underwear}
\docType{data}
\title{South Korean Underwear}
\description{
The dataset comes from a large survey conducted by a South Korean underwear manufacturer in 1997 (Yang, 1997). 664 South Korean consumers were asked to provide responses for three multiple-choice items: preferred brand of underwear (8 brands), attributes when considering a brand of underwear to purchase (15 attributes) and consumer age (6 levels).
}
\usage{data(underwear)}
\format{
  A data frame with 664 observations on the following 3 variables.
  \describe{
    \item{\code{brand}}{categorical}
    \item{\code{atts}}{categorical}
    \item{\code{age}}{categorical}
  }
}
\source{Yang 1997}
\references{
Yang, B. (1997). A consumer research report on underwear in South Korea. Unpublished report, Chung-Ang University, Seoul, South Korea.
}
\examples{
data(underwear)
}
\keyword{datasets}
